Day 17, false start
This version was oversimplified. The "narrow chamber" is modeled as just 7 numbers, the highest point of rock in each column, but there are cavities below these top points where shapes can fit and continue falling. This version gives the correct result for the example, but not for the puzzle input.